// src/lib/types/TailwindGradientOrientation.ts
export type TailwindLinearGradientOrientation =
  | 0
  | 45
  | 90
  | 135
  | 180
  | 225
  | 270
  | 315;

  export type TailwindRadialGradientOrientation =
  | 315
  | "circle at top left"
  | "circle at top"
  | "circle at top right"
  | "circle at left"
  | "circle at center"
  | "circle at right"
  | "circle at bottom left"
  | "circle at bottom"
  | "circle at bottom right"